backend/requirements.txt
### backend/requirements.txt

# Core Framework
fastapi==0.118.3
uvicorn==0.37.0
starlette==0.48.0
pydantic==2.12.0
pydantic_core==2.41.1
typing_extensions==4.15.0

# Database & ORM
SQLAlchemy==2.0.43
alembic==1.16.5
psycopg==3.2.10
psycopg-binary==3.2.10

# Authentication & Security
python-jose==3.5.0
Authlib==1.6.5
passlib==1.7.4
bcrypt==3.2.0
cryptography==46.0.2
email-validator==2.3.0

# API & HTTP
httpx==0.28.1
httpcore==1.0.9
requests==2.32.5
h11==0.16.0
httptools==0.7.1

# Utilities
python-dotenv==1.1.1
python-multipart==0.0.20
python-dateutil==2.9.0.post0
click==8.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
wrapt==1.17.3

# Rate Limiting & Admin
slowapi==0.1.9
sqladmin==0.21.0
WTForms==3.1.2
limits==5.6.0

# Performance
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
greenlet==3.2.4

# Dependencies Management
packaging==25.0
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
Deprecated==1.2.18
dnspython==2.8.0
ecdsa==0.19.1
idna==3.10
Mako==1.3.10
pyasn1==0.6.1
pycparser==2.23
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.2
urllib3==2.5.0

# XML Processing (CBAM Reports)
lxml==4.9.3

# Celery & Redis (Asenkron Görevler)
celery==5.3.0
redis==5.0.0

# Email Service (SendGrid)
sendgrid==6.10.0

# OCR & Invoice Processing (Modül 2.2)
google-cloud-vision==3.6.0
pdf2image==1.16.3
pytesseract==0.3.10
pillow==11.0.0


# Dev Dependencies
pytest==8.0.0
pytest-asyncio==0.23.5
ruff==0.3.0
